requirements.txt
ai4bharat-transliteration==1.1.3
aiohttp==3.8.4
aiosignal==1.3.1
aksharamukha==2.0.4
alabaster==0.7.16
antlr4-python3-runtime==4.8
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.14.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bitarray==2.9.2
blinker==1.7.0
blis==0.7.8
bs4==0.0.1
catalogue==2.0.7
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
cryptography==37.0.2
ctranslate2==2.17.0
cymem==2.0.6
Cython==3.0.8
Deprecated==1.2.13
docutils==0.20.1
fairseq==0.12.2
filelock==3.13.1
Flask==3.0.1
Flask-Cors==4.0.0
fonttools==4.33.3
frozenlist==1.3.3
fsspec==2023.12.2
gevent==23.9.1
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.20.3
hydra-core==1.0.7
hyperframe==5.2.0
icu==0.0.1
idna==2.10
imagesize==1.4.1
importlib-metadata==7.0.1
indic-nlp-library==0.92
itsdangerous==2.1.2
jaconv==0.3
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
language-data==1.1
libretranslatepy==2.1.1
lxml==4.9.1
lyricsgenius==3.0.1
marisa-trie==0.7.7
MarkupSafe==2.1.1
mock==5.1.0
Morfessor==2.0.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.7
networkx==3.2.1
numpy==1.23.0
omegaconf==2.0.6
openai==0.27.4
outcome==1.2.0
packaging==21.3
pandas==2.2.0
pathy==0.6.2
portalocker==2.8.2
preshed==3.0.6
progressbar2==4.3.2
protobuf==4.21.2
pyarrow==15.0.0
pycld2==0.41
pycparser==2.21
pydantic==1.8.2
pydload==1.0.9
Pygments==2.17.2
PyICU==2.9
pykakasi==2.2.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
rfc3986==1.5.0
roman==3.3
sacrebleu==2.4.0
sacremoses==0.1.1
safetensors==0.4.2
scandir==1.10.0
selenium==4.3.0
sentencepiece==0.1.96
sh==2.0.3
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
Sphinx==7.2.6
sphinx-argparse==0.4.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
srsly==2.4.3
stanza==1.1.1
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
thinc==8.0.17
tokenizers==0.15.1
toml==0.10.2
torch==2.2.0
torchaudio==2.2.0
tqdm==4.64.0
transformers==4.37.2
translate==3.6.1
trio==0.21.0
trio-websocket==0.9.2
typer==0.4.2
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
unicodedata2==14.0.0
urduhack==0.1.4
urllib3==1.26.9
wasabi==0.9.1
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.1.0
yarl==1.9.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
